fix: use client for personas page
